bridge/requirements.txt
fastapi
uvicorn[standard]
watchdog
pydantic
psycopg2-binary
tomli
